+++ packages/terra-core-docs/CHANGELOG.md
@@ -20,6 +20,9 @@
   * Updated broken links in terra-form-select and action-footer.
   * linter fixes.
 
+* Changed
+  * Added test for programmatic focus of hyperlink.
+
 ## 1.13.6 - (May 3, 2022)
 
 * Changed