+++ packages/terra-hyperlink/CHANGELOG.md
@@ -6,6 +6,9 @@
   * `isDisabled` and `isUnderlineHidden` props have been marked as private.
   * Updated jest snapshots.
   
+* Fixed
+  * Fixed focus box bug for hyperlink.
+
 ## 2.51.4 - (May 3, 2022)
 
 * Changed